Task 1,2,3

+++ conFusionServer/routes/dishRouter.js
@@ -20,6 +20,11 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser,(req, res, next) => {
+    if(!authenticate.verifyAdmin(req)) {
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        return next(err);
+    }
     Dishes.create(req.body)
     .then((dish) => {
         console.log('Dish Created ', dish);
@@ -30,10 +35,20 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .put(authenticate.verifyUser,(req, res, next) => {
+    if(!authenticate.verifyAdmin(req)) {
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        return next(err);
+    }
     res.statusCode = 403;
     res.end('PUT operation not supported on /dishes');
 })
 .delete(authenticate.verifyUser,(req, res, next) => {
+    if(!authenticate.verifyAdmin(req)) {
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        return next(err);
+    }
     Dishes.remove({})
     .then((resp) => {
         res.statusCode = 200;
@@ -55,10 +70,20 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser,(req, res, next) => {
+    if(!authenticate.verifyAdmin(req)) {
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        return next(err);
+    }
     res.statusCode = 403;
     res.end('POST operation not supported on /dishes/'+ req.params.dishId);
 })
 .put(authenticate.verifyUser,(req, res, next) => {
+    if(!authenticate.verifyAdmin(req)) {
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        return next(err);
+    }
     Dishes.findByIdAndUpdate(req.params.dishId, {
         $set: req.body
     }, { new: true })
@@ -70,6 +95,11 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser,(req, res, next) => {
+    if(!authenticate.verifyAdmin(req)) {
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        return next(err);
+    }
     Dishes.findByIdAndRemove(req.params.dishId)
     .then((resp) => {
         res.statusCode = 200;
@@ -98,6 +128,11 @@ dishRouter.route('/:dishId/comments')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req, res, next) => {
+    if(!authenticate.verifyAdmin(req)) {
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        return next(err);
+    }
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if (dish != null) {
@@ -123,11 +158,21 @@ dishRouter.route('/:dishId/comments')
     .catch((err) => next(err));
 })
 .put(authenticate.verifyUser,(req, res, next) => {
+    if(!authenticate.verifyAdmin(req)) {
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        return next(err);
+    }
     res.statusCode = 403;
     res.end('PUT operation not supported on /dishes/'
         + req.params.dishId + '/comments');
 })
 .delete(authenticate.verifyUser,(req, res, next) => {
+    if(!authenticate.verifyAdmin(req)) {
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        return next(err);
+    }
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if (dish != null) {
@@ -213,6 +258,11 @@ dishRouter.route('/:dishId/comments/:commentId')
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+    if(!authenticate.verifyAdmin(req)) {
+        var err = new Error('You are not authorized to perform this operation!');
+        err.status = 403;
+        return next(err);
+    }
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if (dish != null && dish.comments.id(req.params.commentId) != null) {
